# Blog Project（Frontend）/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BinSpace API
  version: 2.0.0-draft.1
  description: |
    BinSpace 当前 42 个接口基线与下一阶段功能 2～8 契约。
    认证使用 Spring Security、Bearer JWT 和 Redis Token Session。
    新增接口在后端实现前仅代表约定，不代表当前环境已经可调用。
servers:
  - url: http://localhost:8080
    description: Local Spring Boot
...
paths:
...
  /api/archive:
    get:
      summary: 获取公开归档
      description: 按时间返回 TECH/MOMENT 混合归档；无请求体。
      tags: [Archive]
      operationId: getPublicArchive
      parameters:
        - { $ref: '#/components/parameters/YearQuery' }
        - { $ref: '#/components/parameters/PostTypeQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页归档
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
...
  /api/categories:
    get:
      summary: 获取公开分类
      description: 只返回 active 分类；used=true 时仅返回被公开 TECH 使用的分类；无请求体。
      tags: [Categories]
      operationId: listPublicCategories
      parameters:
        - { $ref: '#/components/parameters/UsedQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页分类
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
  /api/owner/categories:
    get:
      summary: 获取 Owner 分类列表
      description: 支持 active/disabled/all 管理视图；无请求体。
      tags: [Categories]
      operationId: listOwnerCategories
      parameters:
        - { $ref: '#/components/parameters/ActiveStatusQuery' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页分类管理数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryAdminPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    post:
      summary: 创建分类
      description: slug 为空时服务端生成；name/slug 必须唯一。
      tags: [Categories]
      operationId: createCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryDTO' }
      responses:
        '201':
          description: 创建后的分类
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryAdminResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/categories/{id}:
    put:
      summary: 更新分类
      description: 使用 category.version 乐观锁。
      tags: [Categories]
      operationId: updateCategory
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategoryDTO' }
      responses:
        '200':
          description: 更新后的分类
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryAdminResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
    delete:
      summary: 停用分类
      description: 逻辑删除，不清空历史文章 category_id；无请求体。
      tags: [Categories]
      operationId: deleteCategory
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: 停用成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
...
  /api/posts/{postId}/comments:
    get:
      summary: 获取 Post 评论
      description: Comment 必须属于指定公开 Post；无请求体。
      tags: [Comments]
      operationId: listPostComments
      parameters:
        - { $ref: '#/components/parameters/PostIdPath' }
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页评论及 Owner 回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
    post:
      summary: 发表评论
      description: 登录 Visitor 或 Owner 对公开 Post 发表评论；作者身份取自 JWT，不接受客户端作者字段。
      tags: [Visitor Interaction]
      operationId: createComment
      parameters:
        - { $ref: '#/components/parameters/PostIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommentDTO' }
      responses:
        '201':
          description: 评论创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/comments/{commentId}/reply:
    post:
      summary: 回复评论
      description: 每条顶层评论最多一条 Owner 回复；新建返回 201，恢复已删除回复返回 200。
      tags: [Comments]
      operationId: replyToComment
      parameters:
        - { $ref: '#/components/parameters/CommentIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplyDTO' }
      responses:
        '200':
          description: 恢复并更新已删除回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '201':
          description: 新建回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/comments/{commentId}:
    delete:
      summary: 删除评论
      description: 删除顶层评论时软删 Owner 回复并维护 post.comment_count；无请求体。
      tags: [Comments]
      operationId: deleteComment
      parameters:
        - { $ref: '#/components/parameters/CommentIdPath' }
      responses:
        '204': { description: 删除成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/guestbook:
    get:
      summary: 获取空间留言
      description: Guestbook 属于整个空间，不属于 Post；首页预览使用 page=1&pageSize=3；无请求体。
      tags: [Guestbook]
      operationId: listGuestbookEntries
      parameters:
        - { $ref: '#/components/parameters/PageQuery' }
        - { $ref: '#/components/parameters/PageSizeQuery' }
      responses:
        '200':
          description: 分页留言及 Owner 回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestbookPageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      security: []
    post:
      summary: 发布空间留言
      description: 登录 Visitor 或 Owner 发布顶层留言；Guestbook 不接受 postId。
      tags: [Visitor Interaction]
      operationId: createGuestbookEntry
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGuestbookDTO' }
      responses:
        '201':
          description: 留言创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestbookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/guestbook/{entryId}/reply:
    post:
      summary: 回复空间留言
      description: 每条顶层留言最多一条 Owner 回复；新建返回 201，恢复已删除回复返回 200。
      tags: [Guestbook]
      operationId: replyToGuestbookEntry
      parameters:
        - { $ref: '#/components/parameters/EntryIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplyDTO' }
      responses:
        '200':
          description: 恢复并更新已删除回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestbookResponse' }
        '201':
          description: 新建回复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestbookResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
  /api/owner/guestbook/{entryId}:
    delete:
      summary: 删除空间留言
      description: 软删顶层留言及其 Owner 回复；无请求体。
      tags: [Guestbook]
      operationId: deleteGuestbookEntry
      parameters:
        - { $ref: '#/components/parameters/EntryIdPath' }
      responses:
        '204': { description: 删除成功，无响应体 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录/注册接口返回的 JWT Access Token。Visitor 与 Owner 受保护请求均使用 Authorization: Bearer <token>。"
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
    PostIdPath:
      name: postId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
    CommentIdPath:
      name: commentId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
    EntryIdPath:
      name: entryId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BigIntId' }
...
    VersionQuery:
      name: version
      in: query
      required: true
      schema: { type: integer, minimum: 0 }
      example: 3
    PostTypeQuery:
      name: type
      in: query
      required: false
      schema: { type: string, enum: [ALL, TECH, MOMENT], default: ALL }
...
    YearQuery:
      name: year
      in: query
      required: false
      schema: { type: integer, minimum: 1970, maximum: 9999 }
      example: 2026
...
    UsedQuery:
      name: used
      in: query
      required: false
      schema: { type: boolean }
      description: true 时仅返回被公开 TECH 使用的项目。
    ActiveStatusQuery:
      name: status
      in: query
      required: false
      schema: { type: string, enum: [ACTIVE, DISABLED, ALL], default: ALL }
...
  responses:
    BadRequest:
      description: 请求参数或 Bean Validation 失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: BAD_REQUEST, message: 请求参数不正确, data: { title: 标题不能为空 }, traceId: 01JEXAMPLE }
    Unauthorized:
      description: 未携带 Token、JWT 无效/过期，或 Redis Token Session 已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: AUTH_REQUIRED, message: 请先登录, data: null, traceId: 01JEXAMPLE }
...
    Forbidden:
      description: 当前 JWT 身份缺少接口要求的角色或权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: FORBIDDEN, message: 无权执行此操作, data: null, traceId: 01JEXAMPLE }
    NotFound:
      description: 资源不存在、已删除或不可见
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_NOT_FOUND, message: 资源不存在, data: null, traceId: 01JEXAMPLE }
    VersionConflict:
      description: 乐观锁版本冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: VERSION_CONFLICT, message: 数据已被修改, data: null, traceId: 01JEXAMPLE }
    ResourceConflict:
      description: 唯一键、重复回复或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: RESOURCE_CONFLICT, message: 资源状态冲突, data: null, traceId: 01JEXAMPLE }
    Conflict:
      description: 乐观锁、唯一键或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            version:
              value: { code: VERSION_CONFLICT, message: 数据已被修改, data: null, traceId: 01JEXAMPLE }
            resource:
              value: { code: RESOURCE_CONFLICT, message: 资源状态冲突, data: null, traceId: 01JEXAMPLE }
    InternalError:
      description: 未处理的服务端错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: INTERNAL_ERROR, message: 服务暂时不可用, data: null, traceId: 01JEXAMPLE }
...
  schemas:
    NullOnly:
      type: object
      nullable: true
      enum: [null]
      description: OpenAPI 3.0 中供 anyOf 复用的 null 分支。
    BigIntId:
      type: string
      pattern: '^[0-9]+$'
      description: MySQL BIGINT 的十进制字符串，禁止在前端转换为 Number。
      example: '1909508401234567168'
    ApiSuccessBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string, enum: [OK], example: OK }
        message: { type: string, example: success }
        traceId: { type: string, example: 01JEXAMPLE }
    ErrorResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, AUTH_REQUIRED, AUTH_FAILED, FORBIDDEN, RESOURCE_NOT_FOUND, VERSION_CONFLICT, RESOURCE_CONFLICT, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, RATE_LIMITED, UPSTREAM_ERROR, INTERNAL_ERROR]
        message: { type: string }
        data:
          type: object
          nullable: true
          additionalProperties: true
        traceId: { type: string }
...
    CategoryVO:
      type: object
      required: [id, name, slug, description, active]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: Redis }
        slug: { type: string, example: redis }
        description: { type: string, nullable: true, example: Redis 实践记录 }
        active: { type: boolean, example: true }
...
    ArchiveItemVO:
      type: object
      required: [id, type, slug, label, publishedAt]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        type: { type: string, enum: [TECH, MOMENT] }
        slug: { type: string, nullable: true }
        label: { type: string, example: Redis + Lua 如何解决库存超卖 }
        publishedAt: { type: string, format: date-time }
    ArchiveMonthVO:
      type: object
      required: [month, items]
      properties:
        month: { type: integer, minimum: 1, maximum: 12, example: 9 }
        items:
          type: array
          items: { $ref: '#/components/schemas/ArchiveItemVO' }
    ArchiveYearVO:
      type: object
      required: [year, months]
      properties:
        year: { type: integer, example: 2026 }
        months:
          type: array
          items: { $ref: '#/components/schemas/ArchiveMonthVO' }
    ArchiveVO:
      type: object
      required: [groups, page, pageSize, total, totalPages, hasMore]
      properties:
        groups:
          type: array
          items: { $ref: '#/components/schemas/ArchiveYearVO' }
        page: { type: integer, minimum: 1, example: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50, example: 10 }
        total: { type: integer, minimum: 0, example: 16 }
        totalPages: { type: integer, minimum: 0, example: 2 }
        hasMore: { type: boolean, example: true }
    ArchiveResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/ArchiveVO' } }
    CategoryAdminVO:
      allOf:
        - { $ref: '#/components/schemas/CategoryVO' }
        - type: object
          required: [sortOrder, version, updatedAt]
          properties:
            sortOrder: { type: integer, example: 10 }
            version: { type: integer, minimum: 0, example: 0 }
            updatedAt: { type: string, format: date-time }
...
    CreateCategoryDTO:
      type: object
      required: [name, slug, description, sortOrder]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, nullable: true, maxLength: 80 }
        description: { type: string, nullable: true, maxLength: 255 }
        sortOrder: { type: integer }
      example: { name: Redis, slug: redis, description: Redis 实践记录, sortOrder: 10 }
    UpdateCategoryDTO:
      type: object
      required: [name, slug, description, sortOrder, version]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, nullable: true, maxLength: 255 }
        sortOrder: { type: integer }
        version: { type: integer, minimum: 0 }
      example: { name: Redis, slug: redis, description: Redis 与缓存实践, sortOrder: 10, version: 1 }
...
    CategoryPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CategoryVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
    CategoryAdminPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CategoryAdminVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
...
    CategoryPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/CategoryPageVO' } }
    CategoryAdminPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/CategoryAdminPageVO' } }
    CategoryAdminResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/CategoryAdminVO' } }
...
    ReplyDTO:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
      example: { content: 谢谢你的留言。 }
    ReplyAuthorVO:
      type: object
      required: [userId, name, avatar]
      properties:
        userId: { $ref: '#/components/schemas/BigIntId' }
        name: { type: string, example: 玢 }
        avatar: { type: string, nullable: true, example: /media/avatar.jpg }
    ReplyVO:
      type: object
      required: [id, author, content, createdAt]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        author: { $ref: '#/components/schemas/ReplyAuthorVO' }
        content: { type: string, example: 谢谢你的留言。 }
        createdAt: { type: string, format: date-time }
    VisitorAuthorVO:
      type: object
      required: [name, avatar]
      properties:
        name: { type: string, example: 游客 }
        avatar: { type: string, nullable: true }
    CommentVO:
      type: object
      required: [id, postId, author, content, createdAt, ownedByMe, ownerReply]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        postId: { $ref: '#/components/schemas/BigIntId' }
        author: { $ref: '#/components/schemas/VisitorAuthorVO' }
        content: { type: string, example: 这篇文章很有帮助。 }
        createdAt: { type: string, format: date-time }
        ownedByMe: { type: boolean, description: 当前登录用户是否为作者；匿名时为 false, example: false }
        ownerReply:
          anyOf:
            - { $ref: '#/components/schemas/ReplyVO' }
            - { $ref: '#/components/schemas/NullOnly' }
    GuestbookVO:
      type: object
      required: [id, author, content, createdAt, ownedByMe, ownerReply]
      properties:
        id: { $ref: '#/components/schemas/BigIntId' }
        author: { $ref: '#/components/schemas/VisitorAuthorVO' }
        content: { type: string, example: 来逛空间啦。 }
        createdAt: { type: string, format: date-time }
        ownedByMe: { type: boolean, description: 当前登录用户是否为作者；匿名时为 false, example: false }
        ownerReply:
          anyOf:
            - { $ref: '#/components/schemas/ReplyVO' }
            - { $ref: '#/components/schemas/NullOnly' }
    CommentPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CommentVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
    GuestbookPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/GuestbookVO' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasMore: { type: boolean }
    CommentPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/CommentPageVO' } }
    CommentResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/CommentVO' } }
    GuestbookPageResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/GuestbookPageVO' } }
    GuestbookResponse:
      allOf:
        - { $ref: '#/components/schemas/ApiSuccessBase' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/GuestbookVO' } }
...
    CreateCommentDTO:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000, example: 这篇文章很有帮助。 }
...
    CreateGuestbookDTO:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000, example: 来都来了，留下一个脚印。 }
...